+++ deno.json
@@ -17,6 +17,7 @@
   "imports": {
     "src/": "./src/",
     "deps": "./src/deps.ts",
+    "cdktf-deps": "./src/outputs/terraform/cdktf-deps.ts",
     "consts": "./src/constants.ts",
     "test-deps": "./src/tests/deps.ts",
     "get-project-root": "./_here.ts"
