+++ src/deps.ts
@@ -11,6 +11,7 @@ export { delay } from "https://deno.land/std@0.201.0/async/delay.ts";
 export { exists } from "https://deno.land/std@0.201.0/fs/mod.ts";
 export { existsSync } from "https://deno.land/std@0.201.0/fs/mod.ts";
 export { load as loadEnv } from "https://deno.land/std@0.205.0/dotenv/mod.ts";
+export { walkSync } from "https://deno.land/std@0.201.0/fs/walk.ts";
 
 export const YAML = {
   ...yaml,
@@ -97,27 +98,3 @@ export const ccolors = {
     return `${colors.white(errCode)}: ${colors.red(e.toString())}`;
   },
 };
-
-export { Construct } from "npm:constructs";
-export {
-  App,
-  Fn,
-  LocalBackend,
-  TerraformLocal,
-  TerraformOutput,
-  TerraformStack,
-  TerraformVariable,
-} from "npm:cdktf";
-
-// Terraform CDKTF Providers
-import * as CDKTFRandomProvider from "npm:@cdktf/provider-random";
-export const RandomProvider = CDKTFRandomProvider.provider.RandomProvider;
-export const RandomPassword = CDKTFRandomProvider.password.Password;
-
-export * as CDKTFProviderAWS from "npm:@cdktf/provider-aws";
-export * as CDKTFProviderAzure from "npm:@cdktf/provider-azurerm";
-export * as CDKTFProviderGCP from "npm:@cdktf/provider-google";
-export * as CDKTFProviderHelm from "npm:@cdktf/provider-helm";
-export * as CDKTFProviderKubernetes from "npm:@cdktf/provider-kubernetes";
-export * as CDKTFProviderTime from "npm:@cdktf/provider-time";
-export * as CDKTFProviderTls from "npm:@cdktf/provider-tls";
