+++ src/outputs/terraform/CNDICoreTerraformStack.ts
@@ -1,11 +1,10 @@
 import {
+  CDKTFProviderRandom,
   Construct,
-  RandomPassword,
-  RandomProvider,
   TerraformLocal,
   TerraformStack,
   TerraformVariable,
-} from "deps";
+} from "cdktf-deps";
 
 import { useSshRepoAuth } from "src/utils.ts";
 import { CNDIConfig } from "src/types.ts";
@@ -15,6 +14,12 @@ export class CNDITerraformStack extends TerraformStack {
   locals: Record<string, TerraformLocal> = {};
   constructor(scope: Construct, name: string, cndi_config: CNDIConfig) {
     super(scope, name);
+
+    new CDKTFProviderRandom.provider.RandomProvider(
+      this,
+      "cndi_provider_random",
+      {},
+    );
 
     const cndi_project_name = cndi_config.project_name!;
 
@@ -80,13 +85,15 @@ export class CNDITerraformStack extends TerraformStack {
       );
     }
 
-    new RandomProvider(this, "random", {});
-
-    const joinToken = new RandomPassword(this, "cndi_join_token", {
-      length: 32,
-      special: false,
-      upper: false,
-    });
+    const joinToken = new CDKTFProviderRandom.password.Password(
+      this,
+      "cndi_join_token",
+      {
+        length: 32,
+        special: false,
+        upper: false,
+      },
+    );
 
     this.locals.bootstrap_token = new TerraformLocal(
       this,
