+++ src/outputs/terraform/azure/AzureCoreStack.ts
@@ -1,4 +1,4 @@
-import { CDKTFProviderAzure, Construct, TerraformLocal } from "deps";
+import { CDKTFProviderAzure, Construct, TerraformLocal } from "cdktf-deps";
 import { CNDIConfig } from "src/types.ts";
 import { CNDITerraformStack } from "../CNDICoreTerraformStack.ts";
