+++ src/outputs/terraform/gcp/GCPCoreStack.ts
@@ -1,4 +1,4 @@
-import { CDKTFProviderGCP, Construct, TerraformLocal } from "deps";
+import { CDKTFProviderGCP, Construct, TerraformLocal } from "cdktf-deps";
 import { CNDIConfig } from "src/types.ts";
 import { CNDITerraformStack } from "../CNDICoreTerraformStack.ts";
