+++ src/outputs/terraform/gcp/GCPMicrok8sStack.ts
@@ -4,8 +4,9 @@ import {
   CDKTFProviderTime,
   Construct,
   Fn,
+  stageCDKTFStack,
   TerraformOutput,
-} from "deps";
+} from "cdktf-deps";
 
 import {
   DEFAULT_INSTANCE_TYPES,
@@ -15,7 +16,6 @@ import {
 import {
   getCDKTFAppConfig,
   resolveCNDIPorts,
-  stageCDKTFStack,
   useSshRepoAuth,
 } from "src/utils.ts";
 import { CNDIConfig, NodeRole } from "src/types.ts";
