+++ src/utils.ts
@@ -1,5 +1,4 @@
 import {
-  App,
   ccolors,
   deepMerge,
   exists,
@@ -10,12 +9,12 @@ import {
   walk,
   YAML,
 } from "deps";
+
 import { DEFAULT_OPEN_PORTS, error_code_reference } from "consts";
 
 import { CNDIConfig, CNDIPort, NodeRole, TFBlocks } from "src/types.ts";
 
 import emitTelemetryEvent from "src/telemetry/telemetry.ts";
-import { walkSync } from "https://deno.land/std@0.201.0/fs/walk.ts";
 
 const utilsLabel = ccolors.faded("src/utils.ts:");
 
@@ -317,31 +316,6 @@ type CDKTFAppConfig = {
   outdir: string;
 };
 
-async function stageCDKTFStack(app: App) {
-  app.synth();
-  const stagingDirectory = await getStagingDir();
-  const tfHome = path.join(stagingDirectory, "cndi", "terraform");
-  const synthDir = path.join(tfHome, "stacks", "_cndi_stack_");
-  Deno.removeSync(path.join(tfHome, "manifest.json")); // this file is useless and confusing unless using cdktf-cli
-  const synthFiles = walkSync(synthDir, { includeDirs: false });
-  for (const entry of synthFiles) {
-    const destinationAbsPath = entry.path.replace(synthDir, tfHome);
-    if (entry.path.endsWith("cdk.tf.json")) {
-      let jsonStr = await Deno.readTextFile(entry.path);
-      const cdktfObj = JSON.parse(jsonStr);
-      delete cdktfObj.terraform.backend;
-      jsonStr = getPrettyJSONString(cdktfObj);
-      Deno.writeTextFileSync(
-        destinationAbsPath,
-        jsonStr.replaceAll("_cndi_stack_", "."),
-      );
-      Deno.removeSync(entry.path);
-      continue;
-    }
-    Deno.renameSync(entry.path, destinationAbsPath);
-  }
-}
-
 async function getCDKTFAppConfig(): Promise<CDKTFAppConfig> {
   const stagingDirectory = await getStagingDir();
   const outdir = path.join(stagingDirectory, "cndi", "terraform");
@@ -564,7 +538,6 @@ export {
   replaceRange,
   resolveCNDIPorts,
   sha256Digest,
-  stageCDKTFStack,
   stageFile,
   useSshRepoAuth,
 };
